+++ cf/test/test_Field.py
@@ -1561,6 +1561,132 @@ class FieldTest(unittest.TestCase):
                 self.assertEqual(g.construct("aux_x").array, 160)
                 self.assertEqual(g.construct("aux_y").array, 3)
 
+        # Halos: monotonic increasing sequence
+        index = [2, 3, 4, 5]
+        indices = f.indices(0, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 4))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [80, 120, 160, 200]).all())
+
+        indices = f.indices(1, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 6))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [40, 80, 120, 160, 200, 240]).all())
+
+        indices = f.indices(999, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 9))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == f.dimension_coordinate("X").array).all())
+
+        # Halos: non-monotonic sequence
+        index = [2, 3, 4, 1]
+        indices = f.indices(0, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 4))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [80, 120, 160, 40]).all())
+
+        indices = f.indices(1, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 6))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [40, 80, 120, 160, 40, 0]).all())
+
+        for halo in (2, 999):
+            indices = f.indices(halo, grid_longitude=index)
+            g = f[indices]
+            self.assertEqual(g.shape, (1, 10, 7))
+            x = g.dimension_coordinate("X").array
+            self.assertTrue((x == [0, 40, 80, 120, 160, 40, 0]).all())
+
+        # Halos: cyclic slice increasing
+        for index in (cf.wi(70, 200), slice(2, 6)):
+            indices = f.indices(0, grid_longitude=index)
+            g = f[indices]
+            self.assertEqual(g.shape, (1, 10, 4))
+            x = g.dimension_coordinate("X").array
+            self.assertTrue((x == [80, 120, 160, 200]).all())
+
+            indices = f.indices(1, grid_longitude=index)
+            g = f[indices]
+            self.assertEqual(g.shape, (1, 10, 6))
+            x = g.dimension_coordinate("X").array
+            self.assertTrue((x == [40, 80, 120, 160, 200, 240]).all())
+
+            indices = f.indices(999, grid_longitude=index)
+            g = f[indices]
+            self.assertEqual(g.shape, (1, 10, 9))
+            x = g.dimension_coordinate("X").array
+            self.assertTrue(
+                (x == [-120, -80, -40, 0, 40, 80, 120, 160, 200]).all()
+            )
+
+        # Halos: cyclic slice increasing
+        index = cf.wi(-170, 40)
+        indices = f.indices(0, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 6))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [-160, -120, -80, -40, 0, 40]).all())
+
+        indices = f.indices(1, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 8))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [-200, -160, -120, -80, -40, 0, 40, 80]).all())
+
+        indices = f.indices(2, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 9))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue(
+            (x == [-240, -200, -160, -120, -80, -40, 0, 40, 80]).all()
+        )
+
+        # Halos: cyclic slice decreasing
+        index = slice(1, -5, -1)
+        indices = f.indices(0, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 6))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [40, 0, -40, -80, -120, -160]).all())
+
+        indices = f.indices(1, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 8))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue((x == [80, 40, 0, -40, -80, -120, -160, -200]).all())
+
+        indices = f.indices(2, grid_longitude=index)
+        g = f[indices]
+        self.assertEqual(g.shape, (1, 10, 9))
+        x = g.dimension_coordinate("X").array
+        self.assertTrue(
+            (x == [120, 80, 40, 0, -40, -80, -120, -160, -200]).all()
+        )
+
+        # Halos: ancillary masking
+        index = cf.wi(90, 100)
+        indices = f.indices(longitude=index)
+        g = f[indices]
+        self.assertTrue(np.ma.is_masked(g.array))
+
+        for halo in (0, 1):
+            indices = f.indices(halo, longitude=index)
+            g = f[indices]
+            self.assertFalse(np.ma.is_masked(g.array))
+
+        # Test API with 0/1/2 arguments
+        kwargs = {"grid_latitude": [1]}
+        i = f.indices(**kwargs)
+        j = f.indices(0, **kwargs)
+        k = f.indices("compress", 0, **kwargs)
+        self.assertEqual(i, j)
+        self.assertEqual(i, k)
+
         # Subspace has size 0 axis resulting from dask array index
         indices = f.indices(grid_latitude=cf.contains(-23.2))
         with self.assertRaises(IndexError):
@@ -2653,6 +2779,14 @@ class FieldTest(unittest.TestCase):
         g = f.subspace(grid_longitude=20)
         h = f.subspace(grid_longitude=np.float64(20))
         self.assertTrue(g.equals(h))
+
+        # Test API with 0/1/2 arguments
+        kwargs = {"grid_latitude": [1]}
+        i = f.subspace(**kwargs)
+        j = f.subspace(0, **kwargs)
+        k = f.subspace("compress", 0, **kwargs)
+        self.assertEqual(i, j)
+        self.assertEqual(i, k)
 
     def test_Field_auxiliary_to_dimension_to_auxiliary(self):
         f = cf.example_field(0)